major reorganisation - cleanup :)
