+++ src/fake-schema/field/directives/fake/fakers/index.ts
@@ -1,7 +1,5 @@
 export { Fakers, createTypeFakers } from "./Fakers";
-export { Random } from "./Random";
+export { Random } from "../../../util/Random";
 export { TypeFakers } from "./TypeFakers";
-import * as resolve from "./resolve";
-import * as maps from "../maps";
 import * as functions from "./functions";
-export { resolve, maps, functions };
+export { functions };
