--- src/fake-schema/field/directives/fake/fakers/resolve/index.ts
@@ -1,2 +1,0 @@
-export { resolveExample } from "./example";
-export { resolveFake } from "./fake";
