+++ src/fake-schema/field/directives/index.ts
@@ -1,5 +1,5 @@
-import * as example from "./example";
+import * as sample from "./sample";
 import * as fake from "./fake";
 import * as mock from "./mock";
 
-export { example, fake, mock };
+export { sample, fake, mock };
