+++ src/fake-schema/field/resolve/examples/index.ts
@@ -1,14 +1,14 @@
-import { ExampleResolver } from "./ExampleResolver";
+import { ExamplesResolver } from "./ExamplesResolver";
 
 // re-align `typeFieldMap` and `fieldMap` (resolve examples and fakes), using a generic `resultResolver`.
 // Allow `matches` list for both, using `resolveMatches`
-export const resolveExample = ({
+export const resolveExamples = ({
   field,
   type,
   fields = [],
   config = {}
 }: any): any[] => {
-  return new ExampleResolver({
+  return new ExamplesResolver({
     field,
     type,
     fields,
