+++ src/fake-schema/field/resolve/index.ts
@@ -1,0 +1,2 @@
+export { resolveExamples } from "./examples";
+export { resolveFakes } from "./fakes";
